move routes to api package
