+++ .gitignore
@@ -229,3 +229,6 @@ pip-selfcheck.json
 *.code-workspace
 
 # End of https://www.toptal.com/developers/gitignore/api/python,vscode,flask,virtualenv
+
+### Diagramms ###
+*.vpp*
