+++ server/api/async_tasks/email.py
@@ -2,7 +2,7 @@ from flask_mail import Message
 from server import celery, mail
 
 
-@celery.task(name='email.confirm')
+@celery.task(name='mail.confirm')
 def send_message(subject: str, emails: list, html):
     msg = Message(subject,
                   recipients=emails,
