+++ server/api/file/serializer.py
@@ -1,5 +1,5 @@
-from server.file.schema import FileSchema
-from server.file.model import File
+from server.api.file.schema import FileSchema
+from server.api.file.model import File
 
 
 def serialize_file(file: File, many=False):
