+++ server/api/task/serializer.py
@@ -1,5 +1,5 @@
-from server.task.schema import TaskSchema
-from server.task.model import Task
+from server.api.task.schema import TaskSchema
+from server.api.task.model import Task
 
 
 def serialize_task(task: Task, many=False):
