+++ server/api/user/serializer.py
@@ -1,5 +1,5 @@
-from server.user.schema import UserSchema
-from server.user.model import User
+from server.api.user.schema import UserSchema
+from server.api.user.model import User
 
 
 def serialize_user(user: User, many=False):
