--- server/category/serializer.py
@@ -1,6 +1,0 @@
-from server.category.model import Category
-from server.category.schema import CategorySchema
-
-
-def serialize_category(cat: Category, many=False):
-    return CategorySchema(many=many).dump(cat)
