--- server/file/__init__.py
@@ -1,1 +1,0 @@
-from server.file.controller import file_blueprint
