--- server/task/__init__.py
@@ -1,1 +1,0 @@
-from server.task.controller import task_blueprint
